project/requirements.txt
# Data & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP
transformers>=4.35.0
datasets>=2.14.0
torch>=2.1.0
nltk>=3.8.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Utils
pyyaml>=6.0
python-dotenv>=1.0.0
structlog>=23.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
jupyter>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0
